refactor: split handlers into separate files
- Created src/handlers/ directory with individual files for each handler
- Each handler now has its own dedicated file:
  - removeButton.ts - Remove buttons from content
  - removeSection.ts - Remove sections from content
  - prependContent.ts - Prepend content to files
  - appendContent.ts - Append content to files
  - addContentAtLine.ts - Add content at specific line
  - createNote.ts - Create new notes
- Added index.ts to export all handlers
- Updated handlers.ts to re-export from new structure
- Maintained backward compatibility through folder structure
