+++ src/handlers/appendContent.ts
@@ -1,0 +1,46 @@
+import { App, MarkdownView, Notice, TFile } from "obsidian";
+import templater from "../templater";
+
+export const appendContent = async (
+  app: App,
+  insert: string | TFile,
+  lineEnd: number,
+  isTemplater: boolean,
+): Promise<void> => {
+  const activeView = app.workspace.getActiveViewOfType(MarkdownView);
+  if (activeView) {
+    const file = activeView.file;
+    let content = await app.vault.read(file);
+    const contentArray = content.split("\n");
+    let insertionPoint;
+    if (
+      contentArray[lineEnd + 1] &&
+      contentArray[lineEnd + 1].includes("^button")
+    ) {
+      insertionPoint = lineEnd + 2;
+    } else {
+      insertionPoint = lineEnd + 1;
+    }
+    if (typeof insert === "string") {
+      contentArray.splice(insertionPoint, 0, `${insert}`);
+    } else {
+      if (isTemplater) {
+        const runTemplater = await templater(app, insert, file);
+        const content = await app.vault.read(insert);
+        const processed = await runTemplater(content);
+        contentArray.splice(insertionPoint, 0, `${processed}`);
+      } else {
+        // For core Templates plugin, just use insertTemplate directly
+        // Set cursor to the insertion point and let insertTemplate handle everything
+        activeView.editor.setCursor(insertionPoint);
+        await app.internalPlugins?.plugins["templates"].instance
+          .insertTemplate(insert);
+        return; // Don't modify the file again, insertTemplate already did the work
+      }
+    }
+    content = contentArray.join("\n");
+    await app.vault.modify(file, content);
+  } else {
+    new Notice("There was an issue appending content, please try again", 2000);
+  }
+};
