+++ src/handlers/index.ts
@@ -1,0 +1,6 @@
+export { removeButton } from "./removeButton";
+export { removeSection } from "./removeSection";
+export { prependContent } from "./prependContent";
+export { appendContent } from "./appendContent";
+export { addContentAtLine } from "./addContentAtLine";
+export { createNote } from "./createNote";
